feat: #107 Deck update API refactor - 덱 수정 시 폴더 업데이트 추가 - 불필요 try catch 로직 제거 - attachment exception 처리 수정

+++ src/main/java/com/edio/studywithcard/card/service/CardServiceImpl.java
@@ -180,7 +180,9 @@ public class CardServiceImpl implements CardService {
                 .filter(Objects::nonNull)
                 .collect(Collectors.toList());
 
-        attachmentService.deleteAllAttachments(oldFileKeys);
+        if (!oldFileKeys.isEmpty()) {
+            attachmentService.deleteAllAttachments(oldFileKeys);
+        }
 
         // 파일이 존재하는 항목만 새 파일 업로드 대상으로 처리
         List<AttachmentBulkData> attachmentsToUpload = updateAttachments.stream()
